refactor/#252: 프로젝트 서비스 레이어 기능별로 분리 (#257)
* Resolved merge conflicts and completed rebase

* refactor/#252: 사용하지 않는 txt파일 삭제

* refactor/#252: 프로젝트 서비스 레이어 기능별로 분리 (Command/Query)

* refactor/#252: PagignUtil static으로 변경하여 도입

* refactor/#252: 사용하지 않는 import 제거

* refactor/#252: 사용하지 프로젝트 전체 조회 api 제거

* refactor/#252: 사용하지 프로젝트 전체 조회 api 제거
