+++ src/main/java/inha/git/project/domain/repository/ProjectQueryRepository.java
@@ -41,65 +41,6 @@ public class ProjectQueryRepository {
     private final JPAQueryFactory queryFactory;
 
     /**
-     * 프로젝트 목록 조회
-     *
-     * @param pageable 페이지 정보
-     * @return 프로젝트 페이지
-     */
-    public Page<SearchProjectsResponse> getProjects(Pageable pageable) {
-        QProject project = QProject.project;
-        QUser user = QUser.user;
-        QProjectField projectField = QProjectField.projectField;
-
-        // 프로젝트 목록 조회 쿼리
-        JPAQuery<Project> query = queryFactory
-                .select(project)
-                .from(project)
-                .leftJoin(project.user, user)
-                .leftJoin(project.projectFields, projectField)
-                .where(project.state.eq(Project.State.ACTIVE))
-                .orderBy(project.id.desc())
-                .offset(pageable.getOffset())
-                .limit(pageable.getPageSize());
-
-        // 결과 리스트 및 총 개수 가져오기
-        List<Project> projects = query.fetch();
-        long total = query.fetchCount();
-        // SearchProjectsResponse 변환
-        List<SearchProjectsResponse> content = projects.stream()
-                .map(p -> new SearchProjectsResponse(
-                        p.getId(),
-                        p.getTitle(),
-                        p.getCreatedAt(),
-                        p.getRepoName() != null,
-                        new SearchSemesterResponse(
-                                p.getSemester().getId(),
-                                p.getSemester().getName()),
-                        new SearchCategoryResponse(
-                                p.getCategory().getId(),
-                                p.getCategory().getName()),
-                        p.getSubjectName(),
-                        p.getLikeCount(),
-                        p.getCommentCount(),
-                        p.getIsPublic(),
-                        p.getProjectFields().stream()
-                                .map(f -> new SearchFieldResponse(
-                                        f.getField().getId(),
-                                        f.getField().getName()
-                                ))
-                                .toList(),
-                        new SearchUserResponse(
-                                p.getUser().getId(),
-                                p.getUser().getName(),
-                                mapRoleToPosition(p.getUser().getRole())
-                        )
-
-                ))
-                .toList();
-        return new PageImpl<>(content, pageable, total);
-    }
-
-    /**
      * 특정 유저의 프로젝트 목록 조회
      *
      * @param id       유저 ID
