+++ src/main/java/inha/git/utils/PagingUtils.java
@@ -1,20 +1,39 @@
 package inha.git.utils;
 
 import inha.git.common.exceptions.BaseException;
-import org.springframework.stereotype.Component;
 
 import static inha.git.common.code.status.ErrorStatus.INVALID_PAGE;
-import static inha.git.common.code.status.ErrorStatus.INVALID_SIZE;
+import static inha.git.common.code.status.ErrorStatus.UTILITY_CLASS;
 
 /**
  * 페이징 처리를 위한 유틸리티 클래스입니다.
  * 페이지 번호와 크기의 유효성 검증 및 변환 기능을 제공합니다.
  */
-@Component
 public class PagingUtils {
 
-    private static final int MIN_PAGE = 1;
-    private static final int MIN_SIZE = 1;
+    private static final Integer MIN_PAGE = 1;
+    private static final Integer MIN_SIZE = 1;
+
+    private PagingUtils() {
+        throw new BaseException(UTILITY_CLASS);
+    }
+
+    /**
+     * 페이지 번호와 크기의 유효성을 검증합니다.
+     *
+     * @param page 검증할 페이지 번호
+     * @param size 검증할 페이지 크기
+     * @throws BaseException INVALID_PAGE: 페이지 번호가 최소값보다 작은 경우
+     *                      INVALID_SIZE: 페이지 크기가 최소값보다 작은 경우
+     */
+    public static void validatePage(Integer page, Integer size) {
+        if (page < MIN_PAGE) {
+            throw new BaseException(INVALID_PAGE);
+        }
+        if (size < MIN_SIZE) {
+            throw new BaseException(INVALID_PAGE);
+        }
+    }
 
     /**
      * 페이지 번호의 유효성을 검증합니다.
@@ -22,21 +41,9 @@ public class PagingUtils {
      * @param page 검증할 페이지 번호
      * @throws BaseException INVALID_PAGE: 페이지 번호가 최소값보다 작은 경우
      */
-    public void validatePage(int page) {
+    public static void validatePage(Integer page) {
         if (page < MIN_PAGE) {
             throw new BaseException(INVALID_PAGE);
-        }
-    }
-
-    /**
-     * 페이지 크기의 유효성을 검증합니다.
-     *
-     * @param size 검증할 페이지 크기
-     * @throws BaseException INVALID_SIZE: 페이지 크기가 최소값보다 작은 경우
-     */
-    public void validateSize(int size) {
-        if (size < MIN_SIZE) {
-            throw new BaseException(INVALID_SIZE);
         }
     }
 
@@ -47,17 +54,7 @@ public class PagingUtils {
      * @param page 변환할 페이지 번호
      * @return 변환된 페이지 인덱스
      */
-    public int toPageIndex(int page) {
+    public static Integer toPageIndex(Integer page) {
         return page - 1;
-    }
-
-    /**
-     * 사용자가 입력한 페이지 크기를 실제 크기로 변환합니다.
-     *
-     * @param size 변환할 페이지 크기
-     * @return 변환된 페이지 크기
-     */
-    public int toPageSize(int size) {
-        return size - 1;
     }
 }
